feat: abstract code to improve code quality
move methods that deal with CSV files to CsvUtil class.
